Absorb PictureLayout into grid
