+++ dotcom-rendering/src/components/DecideLines.tsx
@@ -43,18 +43,20 @@ const DottedLines = ({
 export const DecideLines = ({ format, color }: Props) => {
 	const count = format.design === ArticleDesign.Comment ? 8 : 4;
 
-	switch (format.theme) {
-		case Pillar.Sport:
-			return <DottedLines count={count} color={color} />;
-		default:
-			return (
-				<StraightLines
-					cssOverrides={css`
-						display: block;
-					`}
-					count={count}
-					color={color}
-				/>
-			);
+	if (
+		format.theme === Pillar.Sport &&
+		format.design !== ArticleDesign.Picture
+	) {
+		return <DottedLines count={count} color={color} />;
 	}
+
+	return (
+		<StraightLines
+			cssOverrides={css`
+				display: block;
+			`}
+			count={count}
+			color={color}
+		/>
+	);
 };
